+++ google_images_search/cli.py
@@ -17,11 +17,12 @@ def cli(ctx, developer_key, custom_search_cx):
 
 
 IMAGE_TYPES = ('clipart', 'face', 'lineart', 'news', 'photo', )
-IMAGE_SIZES = ('huge', 'icon', 'large', 'medium', 'small', 'xlarge', 'xxlarge')
+IMAGE_SIZES = ('HUGE', 'ICON', 'LARGE', 'MEDIUM', 'SMALL', 'XLARGE', 'XXLARGE')
 FILE_TYPES = ('jpg', 'gif', 'png')
 DOMINANT_COLORS = ('black', 'blue', 'brown', 'gray', 'green', 'pink', 'purple',
                    'teal', 'white', 'yellow')
 SAFE_SEARCH = ('high', 'medium', 'off', )
+USAGE_RIGHTS = ('cc_publicdomain', 'cc_attribute', 'cc_sharealike', 'cc_noncommercial', 'cc_nonderived')
 
 
 @cli.command()
@@ -35,15 +36,18 @@ SAFE_SEARCH = ('high', 'medium', 'off', )
 @click.option('-i', '--imagetype', type=click.Choice(IMAGE_TYPES),
               default='photo', help='Image type')
 @click.option('-s', '--imagesize', type=click.Choice(IMAGE_SIZES),
-              default='large', help='Image size')
+              default='LARGE', help='Image size')
 @click.option('-c', '--dominantcolor', type=click.Choice(DOMINANT_COLORS),
               default='black', help='Dominant color in images')
+@click.option('-r', '--usagerights', type=click.Choice(USAGE_RIGHTS), multiple=True,
+              default=('cc_publicdomain',), help='Usage rights of images')
 @click.option('-d', '--download_path', type=click.Path(dir_okay=True),
               help='Download images')
 @click.option('-w', '--width', help='Image crop width')
 @click.option('-h', '--height', help='Image crop height')
 def search(ctx, query, num, safe, filetype, imagetype,
-           imagesize, dominantcolor, download_path, width, height):
+           imagesize, dominantcolor, usagerights, download_path, width, height):
+    usagerights = '|'.join(usagerights)
 
     search_params = {
         'q': query,
@@ -52,7 +56,8 @@ def search(ctx, query, num, safe, filetype, imagetype,
         'fileType': filetype,
         'imgType': imagetype,
         'imgSize': imagesize,
-        'imgDominantColor': dominantcolor
+        'imgDominantColor': dominantcolor,
+        'rights': usagerights
     }
 
     click.clear()